Add Logstash Configuration directory.
- Adds `_cat/shards` configuration for Logstash.

In the longer term it would be nice to have a script that we can point at a `tar`/`zip` from the support diagnostics, which could then auto-run all configurations, but it's easy enough to run them specifically for now.

It's probably useful to note that a lot of the results from the support-diagnostics script are from the same time, but this is not currently reflected in indexed results.