Remove config directory from diagnostics bundle
The config directory is pulled back for the elasticsearch.yml and
jvm.options yet the configuration values dictated by these files are
already available via API calls. Additonally, the configuration options
in these files might not the ones in effect if command line options were
provided that override values in these files, or if the file was edited
after the instance was started. Therefore, the correct source for these
values is the API calls (which are already included in the diagnostics
bundle) and we no longer have a need to pull back the config directory.

Relates #192