Add watcher_stack back to commercial versions
watcher_stack, the emit_stacktraces=true API call for watcher was not
added to newer versions of the diags tool. When diagnosing a cluster
recently, we wanted to look at these. This will add it back to all major
versions in the event we need to look at it next time.