Consolidate REST Calls
Refactor configuration to start with a set of calls common to all versions and then modify incrementally based on the version being queried. This will also allow it to be used with newer versions that have not had an explicit entry added to the configuration, instead falling back to the last configured version's calls.