Write full non-200 responses to file.
It's been requested that when any kind of response comes back to write it to the target file even if it's a 400 status. I used to do this but had removed it based on a previous request. It appears that more people are OK with having the response in the file than otherwise so I am putting it back. Diverting this to the logs is touchy, and reverting this allows the query execution to be simplified.

Fixes per review

Writing non-200 responses to file and adding close for output streams.