Fix handling of PKI keystore
This commit resolves the following issues with PKI keystores.

- The keystore file is validated to see whether it exists, but the
  check was backwards, and it would fail if the file _did_ exist.
- The error message printed '{}' rather than the path of the file.
- We would throw a NPE if the keystore password wasn't provided.
- The keystore was loaded after the SSL context was already built so
  it was not actually used by the client.

Note: Since f4911c19, the keystore file is now validated even if the
keystore password is not provided. Previously the keystore was
silently ignored if there was no password option. This is change is
arguably a good thing, but it means users who previously might have
had their invalid "--keystore /some/random/file" ignored, now receive
an error, so we need to make sure the tool is well behaved here.