Add Runner to Manifest
This adds the concept of a `runner` to indicate who or what ran the
diagnostic.

It also fixes a few things, such as forcing the UTC date functions
to actually force UTC, as well as printing the ISO-8601 formatted
timezone instead of the RFC 822 ones.

Finally, it fixes a malformed ES action (because it should match
one and only one API per action) and removes some Kibana tests that
did not reflect the current state of output.